# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AnyCompany Users Service API",
    "description": "RESTful API for managing user records with JWT-based authentication via Amazon Cognito.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://{api_id}.execute-api.{region}.amazonaws.com/v1",
      "variables": {
        "api_id": { "default": "YOUR_API_ID" },
        "region": { "default": "us-east-1" }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "CognitoTokenAuthorizer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Cognito-issued JWT. Obtain via POST /auth/login. Pass as Authorization: Bearer <token>."
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": ["refresh_token"],
        "properties": {
          "refresh_token": { "type": "string" }
        }
      },
      "LogoutRequest": {
        "type": "object",
        "required": ["access_token"],
        "properties": {
          "access_token": { "type": "string" }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "id_token": { "type": "string" },
          "refresh_token": { "type": "string" },
          "expires_in": { "type": "integer" }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "required": ["email", "name"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string", "minLength": 1, "maxLength": 128 }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string", "minLength": 1, "maxLength": 128 }
        },
        "description": "At least one field required."
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "user_id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "status": { "type": "string", "enum": ["ACTIVE"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "ListUsersResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserResponse" }
          },
          "count": { "type": "integer" },
          "next_token": { "type": "string", "description": "Pagination cursor; absent on last page." }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "status_code": { "type": "integer" }
        }
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "loginUser",
        "summary": "Authenticate a user",
        "description": "Exchanges email + password for Cognito tokens. No authorization required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful authentication",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "refreshToken",
        "summary": "Refresh access token",
        "description": "Exchanges a refresh token for a new access token. No authorization required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New access token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "401": { "description": "Invalid or expired refresh token" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logoutUser",
        "summary": "Invalidate all tokens for a user",
        "description": "Calls Cognito GlobalSignOut. No authorization required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LogoutRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Successfully logged out" },
          "401": { "description": "Invalid access token" }
        }
      }
    },
    "/users": {
      "get": {
        "operationId": "listUsers",
        "summary": "List all users",
        "security": [{ "CognitoTokenAuthorizer": [] }],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          },
          {
            "name": "next_token",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Pagination cursor from a previous response."
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListUsersResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized — missing or invalid token" }
        }
      },
      "post": {
        "operationId": "createUser",
        "summary": "Create a new user",
        "security": [{ "CognitoTokenAuthorizer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "409": { "description": "Email already exists" },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/users/{user_id}": {
      "parameters": [
        {
          "name": "user_id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" }
        }
      ],
      "get": {
        "operationId": "getUser",
        "summary": "Get a user by ID",
        "security": [{ "CognitoTokenAuthorizer": [] }],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": { "description": "Invalid UUID format" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "operationId": "updateUser",
        "summary": "Update a user",
        "security": [{ "CognitoTokenAuthorizer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": { "description": "No updatable fields or invalid UUID" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "409": { "description": "Email already taken by another user" }
        }
      },
      "delete": {
        "operationId": "deleteUser",
        "summary": "Delete a user",
        "security": [{ "CognitoTokenAuthorizer": [] }],
        "responses": {
          "204": { "description": "User deleted" },
          "400": { "description": "Invalid UUID format" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" }
        }
      }
    }
  }
}